Add Synthea postInstall file
Add another command line argument which will allow the user to install
a version of the Synthea Ingestor patch using a supplied DUZ value as
the installer.

Add information about the DUZ that needs to be supplied.  At this point,
it can be added as a command line value for FOIA based instances, but
not VEHU.