Install Utilities required by Kernel-GTM project
Reworked so yum install is only done once: In Dockerfile for docker, in
bootstarp for other systems. Bootstrap not tested right now; but the
code is there if anybody wants to pick it up.

This is required so that hashing, reverse ip resolution, etc. would
work.

Here is the full list of required utilities according to the Kernel-GTM
project.

GTM/YDB:

    lsof
    ps
    openssl
    shasum
    xxd
    base64
    dig
    ipcs
    rm/del
    mv
    stat
    mkdir
    wget
    file
    gzip
    dos2unix

Cache requires the following:

    openssl
    rm
    mv
    stat
    mkdir
    wget
    file
    gzip
    dos2unix

Tested with VEHU/GT.M and RPMS/Caché.