Bats Integrated Testing as part of Docker Build
- -t Option in autoInstaller.sh to run BATS testing. BATS testing runs
start.sh in the background, waits 5 seconds, runs tests, and then
SIGTERMs start.sh to cause the shutdown, and wait 5 seconds. The wait is
baaaaaaaad (race conditions etc) but I don't know what else to do for
the initial draft of this code.
- DEVZERO EP in KBANTCLN sets up Zero SLAVE device as GT.M makes that
the device in HEREDOCs. Key field is SIGN-ON/SYSTEM DEVICE needs to be
set for the SLAVE device. Technically not correct; but it doesn't
matter. Needed so that HOME^%ZIS won't fail with an error of 'sign on
system device not found'
- Myriad fixes to tests/vista.bats:
-- GTM Tests work for both GTM/YDB now.
-- Fileman was capitalized to FileMan between Fileman versions. Now we
   just look for "File".
-- RPC Broker Output includes some control characters from HOME^%ZIS.
   Account for that in the output.
-- VistALink Test: Remove cloned folder; otherwise, some bash commands
   fail because the folder already exists.
-- VistALink Test: Simplify classpath logic.
-- Do a pretty thorough job of finding access codes and verify codes for
   the VistAlink Tester.

VistALink tester needs more work. It doesn't work on instances of VistA
where the access and verify code are hashed, as the tester tries to
extract the access/verify code from the NEW PERSON file, and they may be
hashed there.

Tested on:
VEHU, VEHU-Plan 6 (all tests pass)
OSEHRA (which just broke this morning for other reasons, so I couldn't
continue testing)
vxVistA (VL doesn't pass due to bad access/verify code)
WorldVistA (VL doesn't pass due to bad access/verify code)