OTF changes broke stuff; fix it
In the transition from Python 2 to Python 3, we are resiving some
scripts. One of the revisions is not to bundle pexpect, but rather
install it from pip. Therefore, we need pip. This commit adds pip to the
CentOS install; plus we slightly change the way we unzip the OTF repo in
order to allow us to use a different repo if we want.