Remove mupip set -null=existing in Octo installer
This sets the VistA region to be able to have data with existing NULL
data. However, it had none to begin with. I think this statement is here
for historical reasons. In any case, we recent developments with Octo,
we now require that VistA data it indexes not have NULL enabled.