YottaDB GUI Start-up Code fix; misc
Several previous commits delt with adding the YottaDB GUI code to make
it functional.

Ever since commit
https://gitlab.com/YottaDB/UI/YDBGUI/-/commit/dcec137912c720b70e0860c2a53953d2dca9b7a3,
the webserver become non-forking.

Unfortunately, the code here assumed that it was forking, and as a
result, blocked on the GUI start-up code, resulting in the
`$basedir/bin/start.sh` never finishing, and as a result, some commands
(like the chmod in that file) never executed, causing permissions
problems when using ssh because the relinkctl files don't have group rw
access.

Other misc changes:
- Rocky Linux 8.6 -> 8.7
- SIGTERM in start.sh did not stop the YDBGUI. Now it does.
- Remove duplicate message "Starting YottaDB GUI process"
- Remove 'Stopping PIP' as that's no longer supported and has been
  supplanted with Octo
- YottaDB GUI now starts on port 8089, as previously was the case,
  before commit
  https://gitlab.com/YottaDB/UI/YDBGUI/-/commit/dcec137912c720b70e0860c2a53953d2dca9b7a3.